// frontend/src/app/api/generated/model/administration-route-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdministrationRouteDto { 
    uid?: string;
    code?: string;
    name?: string;
    description?: string;
    active?: boolean;
}