// frontend/src/app/api/generated/model/medication-administration-view.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MedicationAdministrationView { 
    uid?: string;
    prescriptionUid?: string;
    admissionUid?: string;
    nurseUid?: string;
    routeUid?: string;
    administeredAt?: string;
    doseGiven?: string;
    patientResponse?: string;
    createdAt?: string;
}